refactor(tests): extract find_references.rs (3 tests)
Move find_references_* tests from mod.rs into
engine/tests/find_references.rs.

Tests moved (now at engine::tests::find_references::*):
  find_references_uses_named_realm
  find_references_structured_doc_key_returns_empty_locations
  find_references_structured_doc_off_key_returns_error

Tests: 135 passed / 0 failed / 1 ignored (baseline unchanged).

marky-n1h step 8 (1/5).
